requirements.txt
python-telegram-bot[job-queue]==22.7
SQLAlchemy==2.0.49
python-dotenv==1.1.1
matplotlib==3.10.1
